+++ attributes/default.rb
@@ -31,29 +31,45 @@ default['cacti']['db_file'] = value_for_platform(
   %w(pld) => {
     'default' => '/etc/webapps/cacti/config.php'
   },
-  %w(ubuntu) => {
+  %w(debian ubuntu) => {
     'default' => '/etc/cacti/debian.php'
   }
 )
+default['cacti']['gid'] = nil
 default['cacti']['group'] = value_for_platform(
   %w(pld) => {
     'default' => 'http'
   },
-  %w(centos fedora redhat ubuntu) => {
+  %w(centos fedora redhat) => {
     'default' => 'apache'
+  },
+  %w(debian ubuntu) => {
+    'default' => 'www-data'
   }
 )
+
+default['cacti']['log_file'] = value_for_platform(
+  %w(centos fedora redhat) => {
+    'default' => '/usr/share/cacti/log/cacti.log'
+  },
+  'default' => '/var/log/cacti/cacti.log'
+)
+
+default['cacti']['mysql_provider'] = 'mysql'
 
 default['cacti']['packages'] = value_for_platform(
   %w(centos fedora redhat) => {
     'default' => %w(cacti net-snmp net-snmp-utils perl-LDAP perl-Net-SNMP php-ldap php-mysql php-pecl-apc php-snmp)
+  },
+  %w(debian) => {
+    '~> 7.0' => %w(cacti libsnmp-base libsnmp15 snmp snmpd libnet-ldap-perl libnet-snmp-perl php-net-ldap php5-mysql php-apc php5-snmp),
+    'default' => %w(cacti libsnmp-base libsnmp30 snmp snmpd libnet-ldap-perl libnet-snmp-perl php-net-ldap php5-mysql php-apc php5-snmp)
   },
   %w(pld) => {
     'default' => %w(cacti cacti-setup)
   },
   %w(ubuntu) => {
     %w(12.04 12.10 13.04) => %w(cacti libsnmp-base libsnmp15 snmp snmpd libnet-ldap-perl libnet-snmp-perl php-net-ldap php5-mysql php-apc php5-snmp),
-    '13.10' => %w(cacti libsnmp-base libsnmp30 snmp snmpd libnet-ldap-perl libnet-snmp-perl php-net-ldap php5-mysql php-apc php5-snmp),
     'default' => %w(cacti libsnmp-base libsnmp30 snmp snmpd libnet-ldap-perl libnet-snmp-perl php-net-ldap php5-mysql php-apc php5-snmp)
   }
 )
@@ -64,7 +80,7 @@ default['cacti']['poller_file'] = value_for_platform(
   %w(pld) => {
     'default' => '/usr/sbin/cacti-poller'
   },
-  %w(ubuntu) => {
+  %w(debian ubuntu) => {
     'default' => '/usr/share/cacti/site/poller.php'
   }
 )
@@ -72,11 +88,13 @@ default['cacti']['poller_cmd'] = value_for_platform(
   %w(pld) => {
     'default' => "umask 022; exec #{node['cacti']['poller_file']} >> /var/log/cacti/poller.log 2>&1"
   },
-  %w(centos fedora redhat ubuntu) => {
+  %w(centos debian fedora redhat ubuntu) => {
     'default' => "/usr/bin/php #{node['cacti']['poller_file']} > /dev/null 2>&1"
   }
 )
 
+default['cacti']['sql_dir'] = nil
+default['cacti']['uid'] = nil
 default['cacti']['user'] = 'cacti'
 default['cacti']['version'] = value_for_platform(
   %w(centos fedora redhat) => {
@@ -89,8 +107,13 @@ default['cacti']['version'] = value_for_platform(
   %w(ubuntu) => {
     '12.04' => '0.8.7i',
     %w(12.10 13.04) => '0.8.8a',
-    '13.10' => '0.8.8b',
-    'default' => '0.8.8a'
+    %w(13.10 14.04 14.10 15.04) => '0.8.8b',
+    'default' => '0.8.8f'
+  },
+  %w(debian) => {
+    '~> 7.0' => '0.8.8a',
+    '~> 8.0' => '0.8.8b',
+    'default' => '0.8.8f'
   }
 )
 
@@ -108,26 +131,33 @@ default['cacti']['apache2']['ssl']['certificate_file'] = value_for_platform(
   %w(pld) => {
     'default' => '/etc/httpd/ssl/server.crt'
   },
-  %w(centos fedora redhat ubuntu) => {
+  %w(centos fedora redhat) => {
     'default' => '/etc/pki/tls/certs/localhost.crt'
+  },
+  %w(debian ubuntu) => {
+    'default' => '/etc/ssl/certs/ssl-cert-snakeoil.pem'
   }
 )
 
 default['cacti']['apache2']['ssl']['chain_file'] = ''
 default['cacti']['apache2']['ssl']['enabled'] = true
 default['cacti']['apache2']['ssl']['force'] = false
+
 default['cacti']['apache2']['ssl']['key_file'] = value_for_platform(
   %w(pld) => {
     'default' => '/etc/httpd/ssl/server.key'
   },
-  %w(centos fedora redhat ubuntu) => {
+  %w(centos fedora redhat) => {
     'default' => '/etc/pki/tls/private/localhost.key'
+  },
+  %w(debian ubuntu) => {
+    'default' => '/etc/ssl/private/ssl-cert-snakeoil.key'
   }
 )
 
 # database attributes (override via cacti/server data bag)
 
-default['cacti']['database']['host'] = 'localhost'
+default['cacti']['database']['host'] = '127.0.0.1'
 default['cacti']['database']['name'] = 'cacti'
 default['cacti']['database']['password'] = 'changeit'
 default['cacti']['database']['port'] = 3306
@@ -138,7 +168,8 @@ default['cacti']['database']['user'] = 'cacti'
 
 default['cacti']['rrdtool']['version'] = value_for_platform(
   %w(centos redhat) => {
-    'default' => '1.3'
+    '~> 6.0' => '1.3',
+    'default' => '1.4'
   },
   %w(fedora) => {
     'default' => '1.4'
@@ -147,15 +178,15 @@ default['cacti']['rrdtool']['version'] = value_for_platform(
     '2.0' => '1.2',
     'default' => '1.4'
   },
-  %w(ubuntu) => {
+  %w(debian ubuntu) => {
     'default' => '1.4'
   }
 )
 
 # Spine attributes
 
-default['cacti']['spine']['version'] = node['cacti']['version']
-default['cacti']['spine']['checksum'] = '2226070cd386a4955063a87e99df2fa861988a604a95f39bb8db2a301774b3ee'
+default['cacti']['spine']['checksum'] = nil
+default['cacti']['spine']['enabled'] = false
 default['cacti']['spine']['packages'] = value_for_platform(
   %w(centos fedora redhat) => {
     'default' => %w(net-snmp-devel openssl-devel)
@@ -163,8 +194,8 @@ default['cacti']['spine']['packages'] = value_for_platform(
   %w(pld) => {
     'default' => %w(net-snmp-devel openssl-devel)
   },
-  %w(ubuntu) => {
+  %w(debian ubuntu) => {
     'default' => %w(libsnmp-dev libssl-dev)
   }
 )
-default['cacti']['spine']['url'] = "http://www.cacti.net/downloads/spine/cacti-spine-#{node['cacti']['spine']['version']}.tar.gz"
+default['cacti']['spine']['url'] = nil