+++ providers/graph.rb
@@ -7,9 +7,8 @@ end
 
 def load_current_resource
   # handle name attribute
-  unless @new_resource.graph_template_id
-    @new_resource.graph_template_id @new_resource.name
-  end
+  return unless @new_resource.graph_template_id
+  @new_resource.graph_template_id @new_resource.name
 end
 
 def graph_exists?
@@ -33,6 +32,7 @@ def params
     'host-id' => get_host_id(@new_resource.host_id),
     'graph-type' => @new_resource.graph_type
   }
+  params['graph-title'] = @new_resource.graph_title if @new_resource.graph_title
 
   case @new_resource.graph_type
   when 'cg'