+++ recipes/database.rb
@@ -1,8 +1,32 @@
 settings = Cacti.settings(node)
 
-if settings['database']['host'] == 'localhost'
-  include_recipe 'mysql::server'
-  include_recipe 'database::mysql'
+if settings['database']['host'] == 'localhost' || settings['database']['host'] == '127.0.0.1'
+  mysql2_chef_gem 'default' do
+    provider Chef::Provider::Mysql2ChefGem::Percona if node['cacti']['mysql_provider'] == 'percona'
+    action :install
+  end
+
+  if node['cacti']['mysql_provider'] == 'percona'
+    include_recipe 'percona::client'
+    include_recipe 'percona::server'
+  else
+    mysql_client 'default' do   
+      action :create   
+    end
+
+    mysql_service 'cacti' do
+      port settings['database']['port']
+      version value_for_platform(
+        'ubuntu' => {
+          %w(12.04 12.10 13.04 13.10) => '5.5',
+          'default' => '5.6'
+        },
+        'default' => '5.5'
+      )
+      initial_root_password node['mysql']['server_root_password']
+      action [:create, :start]
+    end
+  end
 
   database_connection = {
     :host => settings['database']['host'],
@@ -11,12 +35,13 @@ if settings['database']['host'] == 'localhost'
       %w(pld) => {
         'default' => 'mysql'
       },
-      %w(centos fedora redhat ubuntu) => {
+      %w(centos debian fedora redhat ubuntu) => {
         'default' => 'root'
       }
     ),
     :password => node['mysql']['server_root_password']
   }
+  mysql_cli_opts = "mysql -h #{database_connection[:host]} -u #{database_connection[:username]} -p#{database_connection[:password]} #{settings['database']['name']}"
 
   mysql_database settings['database']['name'] do
     connection database_connection
@@ -24,21 +49,9 @@ if settings['database']['host'] == 'localhost'
     notifies :run, 'execute[setup_cacti_database]', :immediately
   end
 
-  cacti_sql_dir = value_for_platform(
-    %w(ubuntu) => {
-      'default' => '/usr/share/doc/cacti'
-    },
-    %w(pld) => {
-      'default' => '/usr/share/cacti/sql'
-    },
-    %w(centos fedora redhat) => {
-      'default' => "/usr/share/doc/cacti-#{node['cacti']['version']}"
-    }
-  )
-
   execute 'setup_cacti_database' do
-    cwd cacti_sql_dir
-    command "mysql -u #{database_connection[:username]} -p#{database_connection[:password]} #{settings['database']['name']} < cacti.sql"
+    cwd Cacti.sql_dir(node)
+    command "#{mysql_cli_opts} < cacti.sql"
     action :nothing
   end
 
@@ -49,6 +62,7 @@ if settings['database']['host'] == 'localhost'
     action :drop
   end
 
+  log "Creating database user"
   mysql_database_user settings['database']['user'] do
     connection database_connection
     host '%'
@@ -57,34 +71,20 @@ if settings['database']['host'] == 'localhost'
     action [:create, :grant]
   end
 
-  # Configure base Cacti settings in database
-  mysql_database 'configure_cacti_database_settings' do
-    connection database_connection
-    database_name settings['database']['name']
-
-    cacti_log_path = value_for_platform(
-      %w(pld) => {
-        'default' => '/var/log/cacti/cacti.log'
-      },
-      %w(centos fedora redhat ubuntu) => {
-        'default' => '/usr/share/cacti/log/cacti.log'
-      }
+  template "#{Cacti.sql_dir(node)}/db-settings.sql" do
+    source 'db-settings.sql.erb'
+    owner node['cacti']['user']
+    group node['cacti']['group']
+    mode 00640
+    variables(
+      :settings => settings
     )
+    notifies :run, 'execute[setup_cacti_settings]', :immediately
+  end
 
-    sql <<-SQL
-      INSERT INTO `settings` (`name`,`value`) VALUES ("path_rrdtool","/usr/bin/rrdtool") ON DUPLICATE KEY UPDATE `value`="/usr/bin/rrdtool";
-      INSERT INTO `settings` (`name`,`value`) VALUES ("path_php_binary","/usr/bin/php") ON DUPLICATE KEY UPDATE `value`="/usr/bin/php";
-      INSERT INTO `settings` (`name`,`value`) VALUES ("path_snmpwalk","/usr/bin/snmpwalk") ON DUPLICATE KEY UPDATE `value`="/usr/bin/snmpwalk";
-      INSERT INTO `settings` (`name`,`value`) VALUES ("path_snmpget","/usr/bin/snmpget") ON DUPLICATE KEY UPDATE `value`="/usr/bin/snmpget";
-      INSERT INTO `settings` (`name`,`value`) VALUES ("path_snmpbulkwalk","/usr/bin/snmpbulkwalk") ON DUPLICATE KEY UPDATE `value`="/usr/bin/snmpbulkwalk";
-      INSERT INTO `settings` (`name`,`value`) VALUES ("path_snmpgetnext","/usr/bin/snmpgetnext") ON DUPLICATE KEY UPDATE `value`="/usr/bin/snmpgetnext";
-      INSERT INTO `settings` (`name`,`value`) VALUES ("path_cactilog","#{cacti_log_path}") ON DUPLICATE KEY UPDATE `value`="#{cacti_log_path}";
-      INSERT INTO `settings` (`name`,`value`) VALUES ("snmp_version","net-snmp") ON DUPLICATE KEY UPDATE `value`="net-snmp";
-      INSERT INTO `settings` (`name`,`value`) VALUES ("rrdtool_version","rrd-#{node['cacti']['rrdtool']['version']}.x") ON DUPLICATE KEY UPDATE `value`="rrd-#{node['cacti']['rrdtool']['version']}.x";
-      INSERT INTO `settings` (`name`,`value`) VALUES ("path_webroot","/usr/share/cacti") ON DUPLICATE KEY UPDATE `value`="/usr/share/cacti";
-      UPDATE `user_auth` SET `password`=md5('#{settings['admin']['password']}'), `must_change_password`="" WHERE `username`='admin';
-      UPDATE `version` SET `cacti`="#{node['cacti']['version']}";
-    SQL
-    action :query
+  execute 'setup_cacti_settings' do
+    cwd Cacti.sql_dir(node)
+    command "#{mysql_cli_opts} < db-settings.sql"
+    action :nothing
   end
 end