+++ recipes/spine.rb
@@ -2,24 +2,23 @@ settings = Cacti.settings(node)
 
 # Install Spine dependencies
 include_recipe 'build-essential'
-include_recipe 'mysql::client'
 
 node['cacti']['spine']['packages'].each do |p|
   package p
 end
 
-remote_file "#{Chef::Config[:file_cache_path]}/cacti-spine-#{node['cacti']['spine']['version']}.tar.gz" do
-  source node['cacti']['spine']['url']
-  checksum node['cacti']['spine']['checksum']
+remote_file "#{Chef::Config[:file_cache_path]}/cacti-spine-#{node['cacti']['version']}.tar.gz" do
+  source Cacti.spine_url(node)
+  checksum Cacti.spine_checksum(node)
   mode '0644'
   action :create_if_missing
 end
 
-execute "install_cacti_spine_#{node['cacti']['spine']['version']}" do
+execute "install_cacti_spine_#{node['cacti']['version']}" do
   cwd Chef::Config[:file_cache_path]
   command <<-COMMAND
-    tar -zxf cacti-spine-#{node['cacti']['spine']['version']}.tar.gz
-    cd cacti-spine-#{node['cacti']['spine']['version']}
+    tar -zxf cacti-spine-#{node['cacti']['version']}.tar.gz
+    cd cacti-spine-#{node['cacti']['version']}
     ./configure --build=x86_64-unknown-linux
     make
     chown root:root spine
@@ -38,26 +37,4 @@ template '/etc/spine.conf' do
   variables(
     :database => settings['database']
   )
-end
-
-if settings['database']['host'] == 'localhost'
-  include_recipe 'database::mysql'
-
-  database_connection = {
-    :host => settings['database']['host'],
-    :port => settings['database']['port'],
-    :username => settings['database']['user'],
-    :password => settings['database']['password']
-  }
-
-  # Configure Spine path and set poller type in database
-  mysql_database 'configure_cacti_database_spine_settings' do
-    connection database_connection
-    database_name settings['database']['name']
-    sql <<-SQL
-      INSERT INTO `settings` (`name`,`value`) VALUES ("path_spine","/usr/bin/spine") ON DUPLICATE KEY UPDATE `value`="/usr/bin/spine";
-      INSERT INTO `settings` (`name`,`value`) VALUES ("poller_type", 2) ON DUPLICATE KEY UPDATE `value`=2
-    SQL
-    action :query
-  end
 end